Mapping data types

+++ databricks/koalas/groupby.py
@@ -1019,7 +1019,7 @@ class GroupBy(object, metaclass=ABCMeta):
         0    6
         1    3
         2    4
-        Name: B, dtype: int32
+        Name: B, dtype: int64
 
         >>> def plus_min(x):
         ...     return x + x.min()
@@ -1036,7 +1036,7 @@ class GroupBy(object, metaclass=ABCMeta):
         >>> df.B.groupby(df.A).apply(plus_length).sort_index()
         0    1
         1    2
-        Name: B, dtype: int32
+        Name: B, dtype: int64
 
         The extra arguments to the function can be passed as below.
 
@@ -1045,7 +1045,7 @@ class GroupBy(object, metaclass=ABCMeta):
         >>> df.B.groupby(df.A).apply(calculation, 5, z=10).sort_index()
         0    51
         1    52
-        Name: B, dtype: int32
+        Name: B, dtype: int64
         """
         from pandas.core.base import SelectionMixin
 
@@ -1993,7 +1993,7 @@ class GroupBy(object, metaclass=ABCMeta):
         0    3
         1    4
         2    6
-        Name: B, dtype: int32
+        Name: B, dtype: int64
 
         >>> (df * -1).B.groupby(df.A).transform(abs)
         0    1
